requirements.txt
a2a-sdk==0.2.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
Deprecated==1.2.18
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
sse-starlette==2.3.5
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
wrapt==1.17.2
zipp==3.22.0
